dev_compare_cfg_file

+++ tools/cfg/template_convert.py
@@ -34,8 +34,16 @@ def JinjaRender(module, filename, **kwargs):
 
 
 def render_cfg_file(dst_file_path, template_file, module):
-    with open(dst_file_path, "w") as dst_file:
-        dst_file.write(JinjaRender(module, template_file))
+    if not os.path.isfile(dst_file_path):
+        with open(dst_file_path, "w") as dst_file:
+            dst_file.write(JinjaRender(module, template_file))
+    else:
+        with open(dst_file_path, "r") as dst_file:
+            org_content = dst_file.read()
+        new_content = JinjaRender(module, template_file)
+        if org_content != new_content:
+            with open(dst_file_path, "w") as dst_file:
+                dst_file.write(new_content)
 
 
 def convert_hpp(dst_hpp_path, module=None):
